x/rich/html: refactor
1. All embed objects use the same key "Embed"
2. All html formatting is moved out of basic types and into the Format function

This should allow migrating all the base types (such as Link/Image etc) over to data/
