+++ x/rich/html/font_style_test.go
@@ -5,11 +5,9 @@
 package html_test
 
 import (
-	"fmt"
 	"testing"
 
 	"github.com/dotchain/dot/changes"
-	"github.com/dotchain/dot/x/rich"
 	"github.com/dotchain/dot/x/rich/html"
 )
 
@@ -29,29 +27,5 @@ func TestFontStyleApply(t *testing.T) {
 	c := changes.ChangeSet{replace}
 	if x := normal.Apply(nil, c); x != replace.After {
 		t.Error("Unexpected changeset", x)
-	}
-}
-
-func TestFontStyle(t *testing.T) {
-	styles := []html.FontStyle{
-		html.FontStyleNormal,
-		html.FontStyleItalic,
-		html.FontStyleOblique,
-	}
-
-	for _, style := range styles {
-		t.Run(string(style), func(t *testing.T) {
-			s := rich.NewText("hello", style)
-			expected := fmt.Sprintf(
-				"<span style=\"font-style: %s\">hello</span>",
-				style,
-			)
-			if style == html.FontStyleItalic {
-				expected = "<i>hello</i>"
-			}
-			if x := html.Format(s, nil); x != expected {
-				t.Error("Unexpected", x)
-			}
-		})
 	}
 }
