+++ x/rich/html/font_weight.go
@@ -5,8 +5,6 @@
 package html
 
 import (
-	"fmt"
-
 	"github.com/dotchain/dot/changes"
 )
 
@@ -41,20 +39,4 @@ func (f FontWeight) Apply(ctx changes.Context, c changes.Change) changes.Value {
 		return c.After
 	}
 	return c.(changes.Custom).ApplyTo(ctx, f)
-}
-
-// OpenTag returns the html tag
-func (f FontWeight) OpenTag() string {
-	if f == FontBold {
-		return "<b>"
-	}
-	return fmt.Sprintf("<span style=\"font-weight: %d\">", int(f))
-}
-
-// CloseTag returns the html tag
-func (f FontWeight) CloseTag() string {
-	if f == FontBold {
-		return "</b>"
-	}
-	return "</span>"
 }
