+++ x/rich/html/font_weight_test.go
@@ -5,11 +5,9 @@
 package html_test
 
 import (
-	"fmt"
 	"testing"
 
 	"github.com/dotchain/dot/changes"
-	"github.com/dotchain/dot/x/rich"
 	"github.com/dotchain/dot/x/rich/html"
 )
 
@@ -29,35 +27,5 @@ func TestFontWeightApply(t *testing.T) {
 	c := changes.ChangeSet{replace}
 	if x := html.FontThin.Apply(nil, c); x != replace.After {
 		t.Error("Unexpected changeset", x)
-	}
-}
-
-func TestFontWeight(t *testing.T) {
-	weights := []html.FontWeight{
-		html.FontThin,
-		html.FontExtraLight,
-		html.FontLight,
-		html.FontNormal,
-		html.FontMedium,
-		html.FontSemibold,
-		html.FontBold,
-		html.FontExtraBold,
-		html.FontBlack,
-	}
-	for _, weight := range weights {
-		str := fmt.Sprintf("%d", weight)
-		t.Run(str, func(t *testing.T) {
-			s := rich.NewText("hello", weight)
-			expected := fmt.Sprintf(
-				"<span style=\"font-weight: %d\">hello</span>",
-				weight,
-			)
-			if weight == html.FontBold {
-				expected = "<b>hello</b>"
-			}
-			if x := html.Format(s, nil); x != expected {
-				t.Error("Unexpected", x)
-			}
-		})
 	}
 }
