+++ x/rich/html/heading.go
@@ -5,9 +5,6 @@
 package html
 
 import (
-	"strconv"
-	"strings"
-
 	"github.com/dotchain/dot/changes"
 	"github.com/dotchain/dot/changes/types"
 	"github.com/dotchain/dot/x/rich"
@@ -28,7 +25,7 @@ type Heading struct {
 
 // Name is the key to use with rich.Attrs
 func (h Heading) Name() string {
-	return "Heading"
+	return "Embed"
 }
 
 // Apply implements changes.Value.
@@ -37,36 +34,17 @@ func (h Heading) Apply(ctx changes.Context, c changes.Change) changes.Value {
 }
 
 func (h Heading) get(key interface{}) changes.Value {
-	switch key {
-	case "Level":
+	if key == "Level" {
 		return changes.Atomic{Value: h.Level}
-	case "Text":
-		return h.Text
 	}
-	return changes.Nil
+	return h.Text
 }
 
 func (h Heading) set(key interface{}, v changes.Value) changes.Value {
-	switch key {
-	case "Level":
+	if key == "Level" {
 		h.Level = v.(changes.Atomic).Value.(int)
-	case "Text":
+	} else {
 		h.Text = v.(*rich.Text)
 	}
 	return h
-}
-
-// FormatHTML formats the heading into HTML
-func (h Heading) FormatHTML(b *strings.Builder, f Formatter) {
-	l := h.Level
-	if l < 1 || l > 6 {
-		l = 1
-	}
-	b.WriteString("<h")
-	b.WriteString(strconv.Itoa(l))
-	b.WriteString(">")
-	FormatBuilder(b, h.Text, f)
-	b.WriteString("</h")
-	b.WriteString(strconv.Itoa(l))
-	b.WriteString(">")
 }
