+++ x/rich/html/image.go
@@ -5,10 +5,6 @@
 package html
 
 import (
-	"strings"
-
-	"golang.org/x/net/html"
-
 	"github.com/dotchain/dot/changes"
 	"github.com/dotchain/dot/changes/types"
 	"github.com/dotchain/dot/x/rich"
@@ -27,7 +23,7 @@ type Image struct {
 
 // Name is the key to use with rich.Attrs
 func (i Image) Name() string {
-	return "Image"
+	return "Embed"
 }
 
 // Apply implements changes.Value.
@@ -36,31 +32,17 @@ func (i Image) Apply(ctx changes.Context, c changes.Change) changes.Value {
 }
 
 func (i Image) get(key interface{}) changes.Value {
-	switch key {
-	case "Src":
+	if key == "Src" {
 		return types.S16(i.Src)
-	case "AltText":
-		return types.S16(i.AltText)
 	}
-	return changes.Nil
+	return types.S16(i.AltText)
 }
 
 func (i Image) set(key interface{}, v changes.Value) changes.Value {
-	switch key {
-	case "Src":
+	if key == "Src" {
 		i.Src = string(v.(types.S16))
-	case "AltText":
+	} else {
 		i.AltText = string(v.(types.S16))
 	}
 	return i
-}
-
-// FormatHTML formats the image into HTML
-func (i Image) FormatHTML(b *strings.Builder, f Formatter) {
-	b.WriteString("<img src=\"")
-	b.WriteString(html.EscapeString(i.Src))
-	b.WriteString("\" alt=\"")
-	b.WriteString(html.EscapeString(i.AltText))
-	b.WriteString("\">")
-	b.WriteString("</img>")
 }
